refactor: AggregationScreen抽出でApp.tsxを薄いシェルに
集計画面の左右パネル・state・ロジックをAggregationScreenに移動し、
Appから集計関連のstate 7個とJSX 100行超を削除。
