+++ src/components/screens/aggregation/AggregationScreen.tsx
@@ -1,5 +1,163 @@
+import { useState } from "preact/hooks";
+import CrossConfig from "../../leftPane/CrossConfig";
+import ResultView from "../../rightPane/ResultView";
+import { runDuckDBAggregation } from "../../../lib/duckdbBridge";
+import { buildQuestionDefs, type LayoutMeta } from "../../../lib/layout";
+import { questionKey, type AggResult, type QuestionDef } from "../../../lib/agg/aggregate";
 import { t } from "../../../lib/i18n";
 import { ToggleButton, ToggleGroup } from "../../shared/ToggleButton";
+
+type CsvData = { text: string; fileName: string; headers: string[]; rowCount: number };
+type LayoutData = { json: string; fileName: string; meta: LayoutMeta };
+
+interface AggregationScreenProps {
+  csv: CsvData;
+  layout: LayoutData;
+}
+
+export default function AggregationScreen({ csv, layout }: AggregationScreenProps) {
+  const questions = buildQuestionDefs(csv.headers, layout.meta.colTypes);
+  const questionLabels = layout.meta.questionLabels;
+  const weightCol = Object.entries(layout.meta.colTypes).find(([, t]) => t === "weight")?.[0] ?? "";
+
+  const [crossSelected, setCrossSelected] = useState<Record<string, boolean>>(() => {
+    const sel: Record<string, boolean> = {};
+    questions.forEach((q) => (sel[questionKey(q)] = false));
+    return sel;
+  });
+  const [weightEnabled, setWeightEnabled] = useState(true);
+  const [errorMsg, setErrorMsg] = useState("");
+  const [aggResults, setAggResults] = useState<{
+    results: AggResult[];
+    weightCol: string;
+    rawN: number;
+    layoutMeta: LayoutMeta;
+    crossCols: QuestionDef[];
+  } | null>(null);
+
+  async function runAggregation(): Promise<void> {
+    setErrorMsg("");
+
+    const wCol = weightEnabled ? weightCol : "";
+    const crossCols = questions.filter((q) => crossSelected[questionKey(q)]);
+
+    try {
+      const results = await runDuckDBAggregation({
+        questions,
+        weight_col: wCol,
+        cross_cols: crossCols,
+      });
+      setAggResults({
+        results,
+        weightCol: wCol,
+        rawN: csv.rowCount,
+        layoutMeta: layout.meta,
+        crossCols,
+      });
+    } catch (e) {
+      setErrorMsg(t("error.aggregation", { msg: (e as Error).message }));
+      console.error(e);
+    }
+  }
+
+  return (
+    <>
+      {/* Left Panel */}
+      <div
+        class="flex flex-col overflow-hidden border-r border-border bg-surface max-md:max-h-[50vh] max-md:border-b max-md:border-r-0"
+        role="region"
+        aria-label={t("section.settings")}
+      >
+        {/* Data Summary */}
+        <section class="shrink-0 border-b border-border p-4">
+          <h2 class="mb-3 text-[0.8125rem] font-bold tracking-[0.04em] text-muted">
+            {t("section.summary")}
+          </h2>
+          <div class="text-[0.875rem] leading-relaxed text-text-secondary">
+            <DataSummary
+              csv={{
+                fileName: csv.fileName,
+                rowCount: csv.rowCount,
+                headers: csv.headers,
+              }}
+              layout={{
+                fileName: layout.fileName,
+                colCount: Object.keys(layout.meta.colTypes).length,
+              }}
+            />
+          </div>
+        </section>
+
+        {/* Cross Config */}
+        <section class="flex min-h-0 flex-1 flex-col overflow-hidden border-b border-border p-4">
+          <h2 class="mb-3 text-[0.8125rem] font-bold tracking-[0.04em] text-muted">
+            {t("section.cross")}
+          </h2>
+          <div
+            class="flex min-h-0 flex-1 flex-col gap-1 overflow-y-auto"
+            role="group"
+            aria-label={t("section.cross.label")}
+          >
+            <CrossConfig
+              questions={questions}
+              questionLabels={questionLabels}
+              crossSelected={crossSelected}
+              onToggle={(key, checked) => setCrossSelected((prev) => ({ ...prev, [key]: checked }))}
+            />
+          </div>
+        </section>
+
+        {/* Weight Info */}
+        {weightCol && (
+          <WeightInfo weightCol={weightCol} enabled={weightEnabled} onToggle={setWeightEnabled} />
+        )}
+
+        {/* Error Message */}
+        {errorMsg && (
+          <div
+            class="mx-4 shrink-0 rounded-lg border border-error-border bg-error-bg px-4 py-3 text-sm leading-normal text-danger"
+            role="alert"
+            aria-live="assertive"
+          >
+            {errorMsg}
+          </div>
+        )}
+
+        {/* Run Button */}
+        <button
+          class="mx-4 my-4 min-h-12 w-[calc(100%-32px)] shrink-0 cursor-pointer rounded-lg border-none bg-accent text-base font-bold tracking-[0.02em] text-accent-contrast transition-[background] duration-150 hover:bg-accent-hover active:bg-[var(--color-primary-900)]"
+          onClick={() => runAggregation()}
+        >
+          {t("run.button")}
+        </button>
+      </div>
+
+      {/* Right Panel */}
+      <div class="overflow-y-auto bg-bg p-6" role="region" aria-label={t("section.results")}>
+        {aggResults ? (
+          <div aria-live="polite">
+            <ResultView
+              results={aggResults.results}
+              weightCol={aggResults.weightCol}
+              rawN={aggResults.rawN}
+              layoutMeta={aggResults.layoutMeta}
+              crossCols={aggResults.crossCols}
+            />
+          </div>
+        ) : (
+          <div class="flex h-full flex-col items-center justify-center gap-3 text-muted">
+            <span class="text-[2.5rem]" aria-hidden="true">
+              ⬛
+            </span>
+            <p class="text-base">{t("empty.text")}</p>
+          </div>
+        )}
+      </div>
+    </>
+  );
+}
+
+// --- Sub-components ---
 
 export function DataSummary({
   csv,
